add  前台管理

+++ fqdianshang/src/main/java/com/fq/pojo/Order.java
@@ -27,7 +27,7 @@ public class Order {
         this.customerId = customerId;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
